1.  New meta-data classes for classRepresentation and for inheritance (both extends and subclasses).  More consistent with the green book & opens the way to allowing same class name on different packages.
2.  The methods package no longer copies any objects into the base
library at startup (as a consequence, it no longer uses data.class).
Another consequence is that the actual code for the trace function had
to be moved to the methods package to get the correct version of
class.  The trace in base now constructs and evaluates a call to the
function in methods.