Allow n-dim arrays to be subsetted by an n-column character matrix
The character matrix is converted to the corresponding integer matrix
by matching against the dimnames of the array.  NA values in any row
of the character matrix are propagated to the result.  Unmatched
values result in a subscript out of bounds error.  Empty string "" is
not allowed to match and therefore always results in an error.