Prevent do_listfiles race from causing a write beyond allocation
The list_files code now checks the length of the allocated result
vector and returns when it is full.  This can occur if files are added
to the filesystem between the call to count_files and list_files.

Thanks to Alistair Gee who reported this on the R-devel mailing list.