unique(): compare CHARSXP by address if ASCII and cached
If all CHARSXPs in a given STRSXP have unmarked encoding (ASCII) and are
in the CHARSXP cache, then we compute a hash value based on the address.
Otherwise, the string data is translated to UTF8 and walked to compute
the hash value.  This provides a speed-up for the ASCII + cached case.